parser ctor

+++ src/lang/code_analysis/syntax_token.rs
@@ -1,12 +1,12 @@
 use super::syntax_kind::*;
-#[derive(Debug)]
+#[derive(Debug,Clone)]
 pub struct SyntaxToken {
     pub kind: SyntaxKind,
-    position: i32,
+    position: usize,
     text: String,
 }
 impl SyntaxToken {
-    pub fn new(kind: SyntaxKind, pos: i32, text: &str) -> SyntaxToken {
+    pub fn new(kind: SyntaxKind, pos: usize, text: &str) -> SyntaxToken {
         SyntaxToken {
             kind,
             position: pos,
